Add Rust specific rules v2
I was discussing with mclemencau/mdhudson/slangasek about what we should do
looking forward to rust binaries (like firefox, but less single special
snowflakes) in main.
We excercised a bit based on my current example of mdevctl [1] and out
of this we have efforts to bring the rust toolchain to main as well as
these efforts to define the MIR rules for those.

But the discussion is more wide-reaching as I want some commitment to get the
toolchain itself to be considered supported as well as some guidance how to
best deal with rust in its current state.

At first it seems to be similar to golang, but it is worse. On one hand lib
transitions are often quite hard in rust, see [2].
But more than that I was told that often transitions get stuck for quite
a long time and are way too broken (this might improve and then we want
to make rust less special).

Finally while the concept of dynamic libraries [3] exists, it isn't usable
as effectively only .so files from the very same build [4] are usable reliably
and therefore are like static builds but in multiple files.

In the discussions it was suggested:
> "
> Regarding the overall rust-in-main question: I'm fine with rustc being in
> main provided there's resourcing.  I think packages /using/ rust should only
> be allowed in main via vendoring of their dependencies, because the
> maintenance of rust libraries in Debian is bad right now.
> ...
> When we did the exception for golang, Jamie Strandboge drafted it for the
> Security Team and I think there was some consultation with Foundations /
> Archive Admins.
> ...
> ultimately it's the Security Team that bears the maintenance burden of
> decisions we make about what to put in main, so I would defer to them and
> have them consult whoever they think is relevant.
> "

In addition the extra burden will also be on the owning team and
sustained engineering, so we want to get all their consultation
before we switch to this.

And while this is shockingly different from how everything else works today
it really seems the only doable way *right now*. All the best practise we
are used to (lib only once for re-use and less maintenance) does no more
apply here until rust and the rust ecosystem evolves further.
But while it is shocking at first, compared to all other approaches that came
up so far that at least is realizable in a not too far future.
And we also have to admit that all golang MIRs of the recent years have
eventually went the vendoring-route as well, so that - for now - just
seems to be the more applicable and doable way.

Updates to v1:
- The new rule/template layout has made this much smaller.
- From the first iteration of this we realized that many golang rules
  will just apply to rust the same way - that again made it less noisy
- We also have identified that vendored code in general (not just rust)
  should have an update plan, so we add a rule for that as well.

[1]: https://bugs.launchpad.net/ubuntu/+source/mdevctl/+bug/1942394
[2]: https://github.com/dtolnay/semver-trick
[3]: https://rust-lang.github.io/rfcs/0404-change-prefer-dynamic.html
[4]: https://doc.rust-lang.org/reference/linkage.html

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>