RFC: ensure package built in most recent archive test rebuild
Add a requirement that the source package under consideration successfully built in the most recent archive test rebuild, if the rebuild occurred more recently than the last upload of the package.

 A motivting example where this is an issue is the python-cheroot MIR:

  https://bugs.launchpad.net/ubuntu/+source/python-cheroot/+bug/1930111/

where the last upload for python-cheroot had occurred during impish, but during the jammy first test rebuild, the package failed to build due to some sort of hang that occurs during the build time tests. Having a package that goes into main but won't rebuild successfully is a problem for maintenance.